moved over to continuous flow values

+++ src/ped/Controller.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import java.util.Set;
 
 interface Controller {
-    Tuple<Phase, Map<TurningMovement, Integer>> run();
+    Tuple<Phase, Map<TurningMovement, Double>> run();
 
     void updateTime(double newTime);
 }
